Refactor : current member 서비스 추상화 및 리펙토링
CurrentMember 클래스가 하던 기능을 리펙토링하였습니다.
