+++ src/main/java/nbbang/com/nbbang/domain/chat/dto/message/ChatSendResponseDto.java
@@ -2,18 +2,10 @@ package nbbang.com.nbbang.domain.chat.dto.message;
 
 import lombok.Builder;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nbbang.com.nbbang.domain.chat.domain.Message;
 import nbbang.com.nbbang.domain.chat.domain.MessageType;
-import nbbang.com.nbbang.domain.member.domain.Member;
-import nbbang.com.nbbang.domain.member.dto.MemberDto;
-import nbbang.com.nbbang.domain.member.dto.MemberSimpleResponseDto;
-import nbbang.com.nbbang.domain.party.domain.Party;
-import nbbang.com.nbbang.global.interceptor.CurrentMember;
-import org.springframework.data.annotation.CreatedDate;
 
-import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Data
