+++ src/main/java/nbbang/com/nbbang/domain/member/controller/MemberController.java
@@ -13,7 +13,7 @@ import nbbang.com.nbbang.domain.member.dto.*;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.global.error.ErrorResponse;
 import nbbang.com.nbbang.global.error.exception.CustomIllegalArgumentException;
-import nbbang.com.nbbang.global.interceptor.CurrentMember;
+import nbbang.com.nbbang.global.security.context.CurrentMember;
 import nbbang.com.nbbang.global.response.DefaultResponse;
 import nbbang.com.nbbang.global.response.StatusCode;
 import org.springframework.http.MediaType;
