+++ src/main/java/nbbang/com/nbbang/domain/member/controller/MemberDevelopController.java
@@ -12,7 +12,7 @@ import nbbang.com.nbbang.domain.member.dto.MemberListResponseDto;
 import nbbang.com.nbbang.domain.member.repository.MemberRepository;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.domain.party.service.ManyPartyService;
-import nbbang.com.nbbang.global.interceptor.CurrentMember;
+import nbbang.com.nbbang.global.security.context.CurrentMember;
 import nbbang.com.nbbang.global.response.DefaultResponse;
 import nbbang.com.nbbang.global.response.StatusCode;
 import nbbang.com.nbbang.global.security.SessionMember;
