+++ src/main/java/nbbang/com/nbbang/global/error/GlobalErrorResponseMessage.java
@@ -20,4 +20,6 @@ public class GlobalErrorResponseMessage {
 
     public static final String MAX_FILE_SIZE_ERROR = "허용된 파일 용량 (10 MB) 를 초과하였습니다.";
     public static final String BAD_MULTIPART_FILE_ERROR = "파일 변환에 실패하였습니다. 10MB 이하의 파일이고 올바른 확장자 (png, jpeg, jpg, jfif) 여야 합니다.";
+
+    public static final String AUTHENTICATION_ERROR = "인증 정보를 조회할 수 없습니다.";
 }
