+++ src/main/java/nbbang/com/nbbang/global/socket/StompChannelInterceptor.java
@@ -3,6 +3,7 @@ package nbbang.com.nbbang.global.socket;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import nbbang.com.nbbang.global.security.context.CurrentMember;
 import nbbang.com.nbbang.global.socket.service.ChatRoomService;
 import nbbang.com.nbbang.global.validator.PartyMemberValidator;
 import org.springframework.messaging.Message;
@@ -24,15 +25,11 @@ public class StompChannelInterceptor implements ChannelInterceptor {
 
     private final ChatRoomService chatRoomService;
     private final PartyMemberValidator partyMemberValidator;
+    private final CurrentMember currentMember;
 
     private static final String TOPIC_GLOBAL = "/topic/global";
     private static final String TOPIC_CHATTING = "/topic/chatting";
     private static final String TOPIC_BREAD_BOARD = "/topic/breadBoard";
-
-    private Long memberId(Message<?> message) {
-        Map<String, Object> sessionHeaders = SimpMessageHeaderAccessor.getSessionAttributes(message.getHeaders());
-        return Long.parseLong(sessionHeaders.get("memberId").toString());
-    }
 
     @Override
     public Message<?> preSend(Message<?> message, MessageChannel channel) {
@@ -40,7 +37,7 @@ public class StompChannelInterceptor implements ChannelInterceptor {
         StompHeaderAccessor accessor = StompHeaderAccessor.wrap(message);
         Map<String, Object> attributes = accessor.getSessionAttributes();
 
-        attributes.put("memberId", memberId(message));
+        attributes.put("memberId", currentMember.idFromSocket(message));
         Long memberId = (Long) attributes.get("memberId");
         log.info("[{}] message: {}", accessor.getCommand(), message);
