+++ src/test/java/nbbang/com/nbbang/global/support/controller/ControllerTestParent.java
@@ -3,7 +3,7 @@ package nbbang.com.nbbang.global.support.controller;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.domain.party.service.PartyService;
 import nbbang.com.nbbang.global.config.WebMvcConfig;
-import nbbang.com.nbbang.global.interceptor.CurrentMember;
+import nbbang.com.nbbang.global.security.context.CurrentMember;
 import nbbang.com.nbbang.global.security.CustomOAuth2MemberService;
 import nbbang.com.nbbang.global.security.LogoutService;
 import nbbang.com.nbbang.global.validator.PartyMemberValidator;
